chore(catalyst-ui): extract parseLBPorts to its own module
The react-refresh/only-export-components rule flagged
LoadBalancerFormFields.tsx for exporting both a component and a
helper. Move parseLBPorts to ./parseLBPorts.ts so the form fields
module only exports React components — keeps lint hygiene clean
without changing behaviour. Follow-up to #349.
