+++ products/catalyst/bootstrap/ui/src/components/CrudModals/formFields/LoadBalancerFormFields.tsx
@@ -39,12 +39,5 @@ export function LoadBalancerFormFields({ values, onChange }: LoadBalancerFormFie
   )
 }
 
-/** Parse the CSV port list into the wire-shape [{ port, protocol }]. */
-export function parseLBPorts(csv: string): { port: number; protocol: string }[] {
-  return csv
-    .split(',')
-    .map((s) => s.trim())
-    .filter(Boolean)
-    .map((p) => ({ port: parseInt(p, 10), protocol: 'tcp' }))
-    .filter((p) => Number.isFinite(p.port) && p.port > 0)
-}
+// parseLBPorts lives in ./parseLBPorts.ts so this file only exports
+// React components (per react-refresh/only-export-components).
