+++ products/catalyst/bootstrap/ui/src/components/CrudModals/formFields/index.ts
@@ -9,7 +9,8 @@ export { ClusterFormFields, type ClusterFormValues, type ClusterFormFieldsProps 
 export { VClusterFormFields, type VClusterFormValues, type VClusterFormFieldsProps } from './VClusterFormFields'
 export { NodePoolFormFields, type NodePoolFormValues, type NodePoolFormFieldsProps } from './NodePoolFormFields'
 export { WorkerNodeFormFields, type WorkerNodeFormValues, type WorkerNodeFormFieldsProps } from './WorkerNodeFormFields'
-export { LoadBalancerFormFields, type LoadBalancerFormValues, type LoadBalancerFormFieldsProps, parseLBPorts } from './LoadBalancerFormFields'
+export { LoadBalancerFormFields, type LoadBalancerFormValues, type LoadBalancerFormFieldsProps } from './LoadBalancerFormFields'
+export { parseLBPorts } from './parseLBPorts'
 export { NetworkFormFields, type NetworkFormValues, type NetworkFormFieldsProps } from './NetworkFormFields'
 export { PVCFormFields, type PVCFormValues, type PVCFormFieldsProps } from './PVCFormFields'
 export { BucketFormFields, type BucketFormValues, type BucketFormFieldsProps } from './BucketFormFields'
